+++ src/features/account/components/addresses/Address.tsx
@@ -1,24 +1,23 @@
-import type { Customer } from '@typings/shop/customer';
+// import type { Customer } from '@typings/shop/customer';
 
 interface AddressProps {
-  loggedInUser: Customer | null;
   section: 'shipping' | 'billing';
 }
 
-export default function Address({ loggedInUser, section }: AddressProps) {
-  let country = '';
+export default function Address({ section }: AddressProps) {
+  // let country = '';
 
-  if (loggedInUser && loggedInUser[section]?.country) {
-    switch (loggedInUser[section].country) {
-      case 'GB':
-        country = 'United Kingdom';
-        break;
-    }
-  }
+  // if (loggedInUser && loggedInUser[section]?.country) {
+  //   switch (loggedInUser[section].country) {
+  //     case 'GB':
+  //       country = 'United Kingdom';
+  //       break;
+  //   }
+  // }
 
   return (
     <>
-      {loggedInUser && loggedInUser[section] ? (
+      {/* {loggedInUser && loggedInUser[section] ? (
         <address className="not-italic">
           <p>{loggedInUser[section]?.firstName + ' ' + loggedInUser[section]?.lastName}</p>
           <p>{loggedInUser[section]?.addressLine1}</p>
@@ -29,9 +28,9 @@ export default function Address({ loggedInUser, section }: AddressProps) {
           {country && <p>{country}</p>}
           <p>{loggedInUser[section]?.phone}</p>
         </address>
-      ) : (
-        <p>You have not set up a {section} address yet</p>
-      )}
+      ) : ( */}
+      <p>You have not set up a {section} address yet</p>
+      {/* )} */}
     </>
   );
 }